requirements.txt
django
pymysql
djangorestframework
requests
django-cors-headers